A usability study on Seek.com
According to the Australian Government Department of Jobs and Small Business1, 60% of Australia’s current job opportunities can be found online. With employers turning to online searches as a way of finding the best candidate, how does the modern employee put themselves up for the job? Enter Seek.com.au, The self-proclaimed “No.1 jobs, employment, career and recruitment site”2 in Australia. With multiple websites setting out to attract potential users, has seek ensured they have created a good user experience to retain potential employers and employees? The following is a report on the results of a study on Seek.com.au and its usability in reference to Whitney Queensberry’s 5E’s of usability.