Quantitative Analysis of Australian’s Interest and Engagement with Science
The purpose of this study was to Quantitatively analyse the provided data from Australian participants during National Science Week, from 13 to 21 August 2011[1]. The data was collected in the form of a survey which contained multiple types of data including binary, non-binary and descriptive datasets. The data received was incomplete and as a result some data cleaning was performed to attempt to increase the integrity of the results. As the data is relatively old and erroneous, all results should be taken into perspective and may not be fully indicative of the population of Australia. To analyse the data 2 research questions were provided and another 2 created to comprise the basis of this report. These questions were answered in a both quantitative and qualitative manner and the results formatted by each individual question.